+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - command line args to configure mimir and grafana URLs
+- Support for loading dashboards in organizations
 
 ## [0.10.2] - 2024-12-17