+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2025-01-10
+
 ### Added
 
 - command line args to configure mimir and grafana URLs
@@ -209,7 +211,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.10.2...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/giantswarm/observability-operator/compare/v0.10.2...v0.11.0
 [0.10.2]: https://github.com/giantswarm/observability-operator/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/giantswarm/observability-operator/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/giantswarm/observability-operator/compare/v0.9.1...v0.10.0